Update README.md
Rectangle Area and Perimeter

Using Lab 2.2 as an example, develop a program that will determine the area and perimeter of a rectangle. The length and width should be given as variables (length = 8, width = 3)

To get started, copy your code from Lab 2.2 into the editor panel. Now modify it to find the area and perimeter of a rectangle with a length and width of 8 and 3.

Continue to work on it until you get the following output:

The area of the rectangle is 24
The perimeter of the rectangle is 22